+++ server_modules/settings.js
@@ -1,0 +1,4 @@
+var file = require('./files');
+exports.route = function (app) {
+    file.route(app, 'settings');
+}
