+++ server_modules/stages.js
@@ -1,0 +1,4 @@
+var file = require('./files');
+exports.route = function (app) {
+    file.route(app, 'stages');
+}
